// frontend/src/api/generated/model/createResourceRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CloudOps API
 * Infrastructure resource management, monitoring and asynchronous tasks. Timestamps use UTC ISO-8601. Error responses share ApiError; errors contains field violations.
 * OpenAPI spec version: 0.0.1-SNAPSHOT
 */
import type { CreateResourceRequestStatus } from "./createResourceRequestStatus";
import type { CreateResourceRequestType } from "./createResourceRequestType";
import type { ResourceConfig } from "./resourceConfig";

export interface CreateResourceRequest {
  /**
   * @minLength 0
   * @maxLength 100
   */
  name: string;
  type: CreateResourceRequestType;
  status: CreateResourceRequestStatus;
  /** @exclusiveMinimum 0 */
  organizationId: number;
  config: ResourceConfig;
}