// frontend/src/api/generated/model/resourceResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * CloudOps API
 * Infrastructure resource management, monitoring and asynchronous tasks. Timestamps use UTC ISO-8601. Error responses share ApiError; errors contains field violations.
 * OpenAPI spec version: 0.0.1-SNAPSHOT
 */
import type { ResourceConfig } from "./resourceConfig";
import type { ResourceResponseHealthStatus } from "./resourceResponseHealthStatus";
import type { ResourceResponseStatus } from "./resourceResponseStatus";
import type { ResourceResponseType } from "./resourceResponseType";

export interface ResourceResponse {
  id?: number;
  name?: string;
  type?: ResourceResponseType;
  status?: ResourceResponseStatus;
  healthStatus?: ResourceResponseHealthStatus;
  organizationId?: number;
  config?: ResourceConfig;
  createdAt?: string;
  updatedAt?: string;
}